Fix selector find betting provider on match page
Fixes #248 - Sadly there is not really a test that is easy to add, as the current snapshot techniques do not work for bets. 
The bets are only present on matches that are not yet over and even there change a lot throughout a game. Therefor this PR does not come with tests other than to verify it by hand.